+++ tomcat/src/main/java/org/apache/coyote/http11/controller/LoginController.java
@@ -8,8 +8,8 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.UUID;
-import org.apache.coyote.http11.HttpCookie;
-import org.apache.coyote.http11.HttpCookieConvertor;
+import org.apache.coyote.http11.httprequest.HttpCookie;
+import org.apache.coyote.http11.httprequest.HttpCookieConvertor;
 import org.apache.coyote.http11.HttpStatusCode;
 import org.apache.coyote.http11.Session;
 import org.apache.coyote.http11.httprequest.HttpRequest;
