+++ tomcat/src/main/java/org/apache/coyote/http11/httprequest/HttpCookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.httprequest;
 
 import java.util.HashMap;
 import java.util.Map;
